# contracts/openapi/forgeops.openapi.json
{
  "components": {
    "schemas": {
      "ActionAdapterKind": {
        "enum": [
          "MOCK",
          "DENY_ALL"
        ],
        "title": "ActionAdapterKind",
        "type": "string"
      },
      "BindingRequest": {
        "additionalProperties": false,
        "properties": {
          "bindingRef": {
            "minLength": 1,
            "title": "Bindingref",
            "type": "string"
          }
        },
        "required": [
          "bindingRef"
        ],
        "title": "BindingRequest",
        "type": "object"
      },
...
      "Environment": {
        "enum": [
          "DEV",
          "TEST",
          "INT",
          "PREPROD",
          "PROD"
        ],
        "title": "Environment",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ManifestSubmission": {
        "additionalProperties": false,
        "properties": {
          "artifactPayloadBase64": {
            "title": "Artifactpayloadbase64",
            "type": "string"
          },
          "manifest": {
            "additionalProperties": true,
            "title": "Manifest",
            "type": "object"
          }
        },
        "required": [
          "manifest",
          "artifactPayloadBase64"
        ],
        "title": "ManifestSubmission",
        "type": "object"
      },
...
      "PermissionGrantRequest": {
        "additionalProperties": false,
        "properties": {
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          }
        },
        "required": [
          "permissions"
        ],
        "title": "PermissionGrantRequest",
        "type": "object"
      },
      "ReleaseRequest": {
        "additionalProperties": false,
        "properties": {
          "actionAdapter": {
            "$ref": "#/components/schemas/ActionAdapterKind"
          },
          "environment": {
            "$ref": "#/components/schemas/Environment"
          }
        },
        "required": [
          "environment",
          "actionAdapter"
        ],
        "title": "ReleaseRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local synthetic EPIC-01/02.5 baseline; advisory-only",
    "title": "ForgeOps Platform API",
    "version": "0.2.5"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/audit-events": {
      "get": {
        "operationId": "list_audit_events_v1_audit_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-forgeops-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forgeops-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Audit Events V1 Audit Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit Events",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/v1/platform/status": {
      "get": {
        "operationId": "platform_status_v1_platform_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-forgeops-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forgeops-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Platform Status V1 Platform Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Platform Status",
        "tags": [
          "platform"
        ]
      }
    },
...
    "/v1/scenario-package-installations": {
      "get": {
        "operationId": "list_installations_v1_scenario_package_installations_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-forgeops-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forgeops-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Installations V1 Scenario Package Installations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Installations",
        "tags": [
          "scenario-packages"
        ]
      },
      "post": {
        "operationId": "install_package_v1_scenario_package_installations_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-forgeops-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forgeops-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManifestSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Install Package V1 Scenario Package Installations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Install Package",
        "tags": [
          "scenario-packages"
        ]
      }
    },
    "/v1/scenario-package-installations/{installation_id}/bindings": {
      "post": {
        "operationId": "bind_package_v1_scenario_package_installations__installation_id__bindings_post",
        "parameters": [
          {
            "in": "path",
            "name": "installation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Installation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-forgeops-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forgeops-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Bind Package V1 Scenario Package Installations  Installation Id  Bindings Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bind Package",
        "tags": [
          "scenario-packages"
        ]
      }
    },
    "/v1/scenario-package-installations/{installation_id}/permission-grants": {
      "post": {
        "operationId": "grant_permissions_v1_scenario_package_installations__installation_id__permission_grants_post",
        "parameters": [
          {
            "in": "path",
            "name": "installation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Installation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-forgeops-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forgeops-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Grant Permissions V1 Scenario Package Installations  Installation Id  Permission Grants Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grant Permissions",
        "tags": [
          "scenario-packages"
        ]
      }
    },
    "/v1/scenario-package-installations/{installation_id}/releases": {
      "post": {
        "operationId": "release_package_v1_scenario_package_installations__installation_id__releases_post",
        "parameters": [
          {
            "in": "path",
            "name": "installation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Installation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-forgeops-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forgeops-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Release Package V1 Scenario Package Installations  Installation Id  Releases Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Package",
        "tags": [
          "scenario-packages"
        ]
      }
    },
    "/v1/scenario-package-installations/{installation_id}/releases/{environment}:disable": {
      "post": {
        "operationId": "disable_release_v1_scenario_package_installations__installation_id__releases__environment__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "installation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Installation Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Environment"
            }
          },
          {
            "in": "header",
            "name": "x-forgeops-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forgeops-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Disable Release V1 Scenario Package Installations  Installation Id  Releases  Environment  Disable Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Release",
        "tags": [
          "scenario-packages"
        ]
      }
    },
    "/v1/scenario-package-installations/{installation_id}/releases/{environment}:enable": {
      "post": {
        "operationId": "enable_release_v1_scenario_package_installations__installation_id__releases__environment__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "installation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Installation Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Environment"
            }
          },
          {
            "in": "header",
            "name": "x-forgeops-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forgeops-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Enable Release V1 Scenario Package Installations  Installation Id  Releases  Environment  Enable Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Release",
        "tags": [
          "scenario-packages"
        ]
      }
    },
...
  }
}